Updated the svelte file structure and page routing
